[CT-5-3] refactor: screen 패키지 추가하고 차트게임 화면 이동
