+++ presentation/src/main/java/com/yessorae/presentation/MainActivity.kt
@@ -10,7 +10,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.yessorae.presentation.ui.chartgame.ChartGameScreen
+import com.yessorae.presentation.ui.screen.chartgame.ChartGameScreen
 import com.yessorae.presentation.ui.designsystem.theme.ChartTrainerTheme
 import dagger.hilt.android.AndroidEntryPoint
