+++ presentation/src/main/java/com/yessorae/presentation/ui/screen/chartgame/component/CandleChartUi.kt
@@ -1,4 +1,4 @@
-package com.yessorae.presentation.ui.chartgame.component
+package com.yessorae.presentation.ui.screen.chartgame.component
 
 import android.graphics.Typeface
 import android.text.Layout
@@ -41,7 +41,7 @@ import com.patrykandpatrick.vico.core.common.component.TextComponent
 import com.patrykandpatrick.vico.core.common.copyColor
 import com.patrykandpatrick.vico.core.common.shape.Corner
 import com.patrykandpatrick.vico.core.common.shape.Shape
-import com.yessorae.presentation.ui.chartgame.model.CandleStickChartUi
+import com.yessorae.presentation.ui.screen.chartgame.model.CandleStickChartUi
 import com.yessorae.presentation.ui.designsystem.util.DevicePreviews
 import com.yessorae.presentation.ui.designsystem.util.UiConstants.SHADOW_DY_DP
 import com.yessorae.presentation.ui.designsystem.util.UiConstants.SHADOW_RADIUS_DP
