+++ presentation/src/main/java/com/yessorae/presentation/ui/screen/chartgame/component/order/TradeOrderKeyPad.kt
@@ -1,4 +1,4 @@
-package com.yessorae.presentation.ui.chartgame.component.order
+package com.yessorae.presentation.ui.screen.chartgame.component.order
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -16,7 +16,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import com.yessorae.presentation.R
-import com.yessorae.presentation.ui.chartgame.model.TradeOrderKeyPad
+import com.yessorae.presentation.ui.screen.chartgame.model.TradeOrderKeyPad
 import com.yessorae.presentation.ui.designsystem.component.HorizontalGrid
 import com.yessorae.presentation.ui.designsystem.util.ChartTrainerIcons
 import com.yessorae.presentation.ui.designsystem.util.DevicePreviews
