+++ presentation/src/main/java/com/yessorae/presentation/ui/screen/chartgame/model/ChartGameUserAction.kt
@@ -1,4 +1,4 @@
-package com.yessorae.presentation.ui.chartgame.model
+package com.yessorae.presentation.ui.screen.chartgame.model
 
 sealed interface ChartGameScreenUserAction {
     object ClickNewChartButton : ChartGameScreenUserAction
